refactor(storage): move `Curl*` to a subdirectory
