+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -55,13 +55,13 @@ google_cloud_cpp_storage_hdrs = [
     "internal/compute_engine_util.h",
     "internal/const_buffer.h",
     "internal/crc32c.h",
-    "internal/curl_client.h",
-    "internal/curl_download_request.h",
-    "internal/curl_handle.h",
-    "internal/curl_handle_factory.h",
-    "internal/curl_request.h",
-    "internal/curl_request_builder.h",
-    "internal/curl_wrappers.h",
+    "internal/curl/client.h",
+    "internal/curl/download_request.h",
+    "internal/curl/handle.h",
+    "internal/curl/handle_factory.h",
+    "internal/curl/request.h",
+    "internal/curl/request_builder.h",
+    "internal/curl/wrappers.h",
     "internal/default_object_acl_requests.h",
     "internal/empty_response.h",
     "internal/error_credentials.h",
@@ -179,13 +179,13 @@ google_cloud_cpp_storage_srcs = [
     "internal/compute_engine_util.cc",
     "internal/const_buffer.cc",
     "internal/crc32c.cc",
-    "internal/curl_client.cc",
-    "internal/curl_download_request.cc",
-    "internal/curl_handle.cc",
-    "internal/curl_handle_factory.cc",
-    "internal/curl_request.cc",
-    "internal/curl_request_builder.cc",
-    "internal/curl_wrappers.cc",
+    "internal/curl/client.cc",
+    "internal/curl/download_request.cc",
+    "internal/curl/handle.cc",
+    "internal/curl/handle_factory.cc",
+    "internal/curl/request.cc",
+    "internal/curl/request_builder.cc",
+    "internal/curl/wrappers.cc",
     "internal/default_object_acl_requests.cc",
     "internal/empty_response.cc",
     "internal/error_credentials.cc",
