+++ google/cloud/storage/internal/curl/download_request.cc
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/curl_download_request.h"
+#include "google/cloud/storage/internal/curl/download_request.h"
 #include "google/cloud/storage/internal/binary_data_as_debug_string.h"
-#include "google/cloud/storage/internal/curl_wrappers.h"
+#include "google/cloud/storage/internal/curl/wrappers.h"
 #include "google/cloud/internal/throw_delegate.h"
 #include "google/cloud/log.h"
 #include <curl/multi.h>
