+++ google/cloud/storage/internal/curl/handle_factory.h
