+++ google/cloud/storage/internal/curl/request.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/curl_request.h"
+#include "google/cloud/storage/internal/curl/request.h"
 #include <iostream>
 
 namespace google {
