+++ google/cloud/storage/internal/curl/request_builder.h
@@ -16,9 +16,9 @@
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_CURL_REQUEST_BUILDER_H
 
 #include "google/cloud/storage/internal/complex_option.h"
-#include "google/cloud/storage/internal/curl_download_request.h"
-#include "google/cloud/storage/internal/curl_handle_factory.h"
-#include "google/cloud/storage/internal/curl_request.h"
+#include "google/cloud/storage/internal/curl/download_request.h"
+#include "google/cloud/storage/internal/curl/handle_factory.h"
+#include "google/cloud/storage/internal/curl/request.h"
 #include "google/cloud/storage/version.h"
 #include "google/cloud/storage/well_known_headers.h"
 #include <chrono>
