+++ google/cloud/storage/internal/curl/wrappers.h
