+++ google/cloud/storage/internal/object_acl_requests_test.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/storage/internal/object_acl_requests.h"
-#include "google/cloud/storage/internal/curl_request_builder.h"
+#include "google/cloud/storage/internal/curl/request_builder.h"
 #include "google/cloud/storage/internal/object_access_control_parser.h"
 #include "google/cloud/testing_util/status_matchers.h"
 #include <gmock/gmock.h>
