+++ google/cloud/storage/internal/signed_url_requests.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/storage/internal/signed_url_requests.h"
-#include "google/cloud/storage/internal/curl_handle.h"
+#include "google/cloud/storage/internal/curl/handle.h"
 #include "google/cloud/internal/absl_str_join_quiet.h"
 #include "google/cloud/internal/format_time_point.h"
 #include "google/cloud/internal/sha256_hash.h"
