+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -49,9 +49,9 @@ storage_client_unit_tests = [
     "internal/compute_engine_util_test.cc",
     "internal/const_buffer_test.cc",
     "internal/crc32c_test.cc",
-    "internal/curl_client_test.cc",
-    "internal/curl_download_request_test.cc",
-    "internal/curl_handle_test.cc",
+    "internal/curl/client_test.cc",
+    "internal/curl/download_request_test.cc",
+    "internal/curl/handle_test.cc",
     "internal/default_object_acl_requests_test.cc",
     "internal/generate_message_boundary_test.cc",
     "internal/generic_request_test.cc",
