+++ google/cloud/storage/testing/retry_http_request.h
@@ -15,7 +15,7 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_TESTING_RETRY_HTTP_REQUEST_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_TESTING_RETRY_HTTP_REQUEST_H
 
-#include "google/cloud/storage/internal/curl_request.h"
+#include "google/cloud/storage/internal/curl/request.h"
 #include "google/cloud/storage/internal/http_response.h"
 #include "google/cloud/status_or.h"
 #include <functional>
