+++ google/cloud/storage/tests/curl_request_integration_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/curl_request_builder.h"
+#include "google/cloud/storage/internal/curl/request_builder.h"
 #include "google/cloud/storage/oauth2/anonymous_credentials.h"
 #include "google/cloud/internal/getenv.h"
 #include "google/cloud/log.h"
